feat: add workflow using lookup tables for pulse shaping modes
