+++ packages/essdiffraction/docs/api-reference/index.md
@@ -125,6 +125,10 @@
 
    BeerModMcStasWorkflowKnownPeaks
    BeerModMcStasWorkflow
+   BeerPowderWorkflow
+   BeerPowderWorkflowAnalytical
+   BeerPowderMcStasWorkflow
+   BeerPowderMcStasWorkflowAnalytical
 ```
 
 ### Top-level functions
@@ -150,6 +154,7 @@
    conversions
    data
    io
+   mcstas
    workflow
 ```
