+++ packages/essdiffraction/src/ess/beer/data.py
@@ -31,7 +31,7 @@ _registry = make_registry(
         # Simulations from new model with corrected(?) moderator to detector
         # distance.
         # For correct reduction you need to use
-        # beer.io.mcstas_chopper_delay_from_mode_new_simulations
+        # beer.mcstas.mcstas_chopper_delay_from_mode_new_simulations
         # to obtain the correct WavelengthDefinitionChopperDelay for these files.
         "silicon-mode10-new-model.h5": "md5:98500830f27700fc719634e1acd49944",
         "silicon-mode16-new-model.h5": "md5:393f9287e7d3f97ceedbe64343918413",
